Add XML documentation comments to methods and classes
XML documentation blocks have been added to various methods and classes to improve readability and understanding of the code. This includes summarisations for functionalities, descriptions of parameters and return types which will aid in development and usage of the functions. Part of the additions also includes the creation of '.idea' configuration files.